Conclusão de rotina para atualizaçao de score dos produtos

+++ libs/src/main/java/br/com/hotmart/libs/domain/Product.java
@@ -49,6 +49,6 @@ public class Product extends AbstractEntityBase {
 	private ProductCategory productCategory;
 
 	@Column(name = "score")
-	private int score = 0;
+	private double score = 0;
 
 }
